+++ src/Entities/AccessTokenEntityInterface.php
@@ -1,6 +1,0 @@
-<?php
-namespace DalPraS\OpenId\Server\Entities;
-
-interface AccessTokenEntityInterface extends 
-    \League\OAuth2\Server\Entities\AccessTokenEntityInterface {
-}